// apps/backend/src/model/kanel/public/PlatformConfiguration.ts
// @generated
// This file is automatically generated by Kanel. Do not modify manually.

import type { ServiceInstanceId } from './ServiceInstance';
import type { PlatformContract, PlatformConfigurationStatus } from '../../../__generated__/resolvers-types';

/** Represents the table public.PlatformConfiguration */
export default interface PlatformConfiguration {
  service_instance_id: ServiceInstanceId;

  registerer_id: string;

  platform_id: string;

  tenant_id: string | null;

  tenant_name: string | null;

  platform_url: string;

  platform_title: string;

  platform_version: string | null;

  platform_contract: PlatformContract;

  last_connectivity_check: Date | null;

  token: string;

  status: PlatformConfigurationStatus;
}

/** Represents the initializer for the table public.PlatformConfiguration */
export interface PlatformConfigurationInitializer {
  service_instance_id: ServiceInstanceId;

  registerer_id: string;

  platform_id: string;

  tenant_id?: string | null;

  tenant_name?: string | null;

  platform_url: string;

  platform_title: string;

  platform_version?: string | null;

  platform_contract: PlatformContract;

  last_connectivity_check?: Date | null;

  token: string;

  status?: PlatformConfigurationStatus;
}

/** Represents the mutator for the table public.PlatformConfiguration */
export interface PlatformConfigurationMutator {
  service_instance_id?: ServiceInstanceId;

  registerer_id?: string;

  platform_id?: string;

  tenant_id?: string | null;

  tenant_name?: string | null;

  platform_url?: string;

  platform_title?: string;

  platform_version?: string | null;

  platform_contract?: PlatformContract;

  last_connectivity_check?: Date | null;

  token?: string;

  status?: PlatformConfigurationStatus;
}